[#34] 백버튼 전역으로 빼기
